+++ app/Http/Controllers/Bible/GetBibleController.php
@@ -18,7 +18,6 @@ class GetBibleController extends Controller
 
     public function getBibleChapterText($book_id, $chapter) {
         //TODO: This query needs to change to look up the chapter_id from the chapters table.
-
         $chapter_row = Chapter::where('book_id', $book_id)
                         ->where('book_chapter', $chapter)
                         ->get();
@@ -33,26 +32,10 @@ class GetBibleController extends Controller
             ->get();
 
         return response($verses);
-            // ->cookie('book_id', $book_id)
-            // ->cookie('chapter_id', $chapter);
-        // $cookie = Cookie::make('pizza', 'there is pizza');
-        // return Verse::where('book_id', $book_id)
-        //     ->where('chapter_id', $chapter_id)
-        //     ->get()
-        //     ->cookie($cookie);
     }
 
     public function getChapterById($chapter_id) {
         return Chapter::where('id', $chapter_id)
             ->get();
     }
-
-    // public function test() {
-    //     $_books = [];
-    //     $books = Book::all();
-    //     foreach($books as $book) {
-    //         array_push($_books, $book->book);
-    //     }
-    //     return Book::all();
-    // }
 }